chore(deps): drop sdk-node and unbuild to slim dep tree, fix vulns
Eliminates the Critical protobufjs@7.5.4 advisory (came in via
@opentelemetry/sdk-node → grpc → proto-loader) and a slew of dev
advisories (svgo, rollup, postcss, minimatch, defu<6.1.5, picomatch
<4.0.3, esbuild dev-server) by removing unbuild's transitive deps.

OpenTelemetry:
- Replace @opentelemetry/sdk-node with @opentelemetry/sdk-trace-node
  + @opentelemetry/sdk-trace-base. We only export traces over OTLP/HTTP
  so the metrics/logs/gRPC/instrumentation surface NodeSDK pulls in is
  pure dead weight here.
- src/tracing.ts: NodeSDK → NodeTracerProvider + BatchSpanProcessor.

Build:
- Replace unbuild (rollup + mkdist + cssnano + …) with tsdown
  (rolldown-based, ~16 deps vs unbuild's tree of hundreds).
- tsdown emits .d.mts/.d.cts; package.json `exports` updated to
  per-condition `types` so tsc resolves them correctly under both
  ESM and CJS consumers.
